requirements.txt
textual>=0.40.0
redis>=5.0.0
pymetasploit3>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
openai>=1.0.0
pyyaml>=6.0.0
